+++ Assets/scripts/Entity/Player/PlayerInputHandler.cs
@@ -5,7 +5,6 @@ public class PlayerInputHandler : MonoBehaviour
     private PlayerControls controls;
     private PlayerMovement pm;
     private PlayerAttackHandler pah;
-    [HideInInspector] public PlayerSkillTree pst;
     public static Vector2 mousePos;
 
     private void Awake()
@@ -13,12 +12,10 @@ public class PlayerInputHandler : MonoBehaviour
         controls = new PlayerControls();
         pm = GetComponent<PlayerMovement>();
         pah = GetComponent<PlayerAttackHandler>();
-        pst = GetComponent<PlayerSkillTree>();
     }
     private void OnEnable()
     {
         controls.Player.Enable();
-        controls.UI.Enable();
 
         mousePos = controls.Player.MousePosition.ReadValue<Vector2>();
 
@@ -30,15 +27,7 @@ public class PlayerInputHandler : MonoBehaviour
         controls.Player.Skill.performed += _ => pah.PerformAttack(AttackType.Skill);
         controls.Player.Ultimate.performed += _ => pah.PerformAttack(AttackType.Ultimate);
         controls.Player.Technique.performed += _ => pah.PerformAttack(AttackType.Technique);
-
-        controls.UI.ToggleSkillTree.performed += _ => ToggleSkillTree();
     }
     private void Update() => mousePos = controls.Player.MousePosition.ReadValue<Vector2>();
     private void OnDisable() => controls.Player.Disable();
-
-    private void ToggleSkillTree()
-    {
-        var skillTreeUI = FindAnyObjectByType<SkillTreeUI>(FindObjectsInactive.Include);
-        if (skillTreeUI != null) skillTreeUI.Toggle(gameObject);
-    }
 }