+++ Assets/scripts/Entity/Player/PlayerLevel.cs
@@ -46,14 +46,14 @@ public class PlayerLevel : MonoBehaviour
         cLv = Mathf.RoundToInt(esm.GetStat(StatType.Level));
         esm.AddStat(new(StatType.Level, 1));
 
-        GetComponent<PlayerSkillTree>().skillPoints++;
+        GetComponent<ISkillPointHolder>().AddSkillPoints(1);
 
         esm.AddStat(new(StatType.maxHp, 3));
         esm.AddStat(new(StatType.attack, 1));
         esm.AddStat(new(StatType.Intelligence, 1));
         esm.AddStat(new(StatType.moveSpeed, 0.005f));
 
-        GameController.Instance.SetTitleForDuration("Levelled Up!", 0.4f, 0.2f, 0.2f);
-        GameController.Instance.SetSubtitleForDuration($"{cLv} → {Mathf.RoundToInt(esm.GetStat(StatType.Level))}", 0.4f, 0.2f, 0.2f);
+        IAnnouncer.Current?.SetTitleForDuration("Levelled Up!", 0.4f, 0.2f, 0.2f);
+        IAnnouncer.Current?.SetSubtitleForDuration($"{cLv} → {Mathf.RoundToInt(esm.GetStat(StatType.Level))}", 0.4f, 0.2f, 0.2f);
     }
 }